refactor(webfont-generator): split WOFF2 encoder stages
