+++ packages/webfont-generator/src/formats/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod eot;
 pub(crate) mod woff1;
+pub(crate) mod woff2;
