+++ packages/webfont-generator/src/formats/woff2/fixtures/300.internal-woff2-baseline
