+++ packages/webfont-generator/src/formats/woff2/fixtures/600.internal-woff2-baseline
