+++ packages/webfont-generator/src/sfnt/mod.rs
@@ -3,8 +3,5 @@ mod serialize;
 
 #[cfg(test)]
 pub(crate) use builder::TtfOptions;
-pub(crate) use builder::{
-    CachedCompiledGlyph, Woff2TransformCache, Woff2TransformPayload, build,
-    ttf_options_from_options,
-};
+pub(crate) use builder::{CachedCompiledGlyph, build, ttf_options_from_options};
 pub(crate) use serialize::{SerializedFontTables, SerializedTable};
